Squashed commit
- remove gopool
- replace memory cache impl dgraph-io/ristretto with maypok86/otter
- fix some prefetch bugs
- improve negative caching
- use connpool
- update deps

+++ app/router/config.go
@@ -104,7 +104,7 @@ type LogConfig struct {
 }
 
 type CacheConfig struct {
-	MemSize    int64  `yaml:"mem_size"`
+	MemSize    int    `yaml:"mem_size"`
 	Redis      string `yaml:"redis"`
 	MaximumTTL int    `yaml:"maximum_ttl"`
 	IpMarker   string `yaml:"ip_marker"`
